backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omnitriage-backend"
version = "0.1.0"
description = "Pre-dataset OmniTriage FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "google-genai>=1.0.0",
  "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
